Improves structure of project
