+++ entities/CatFact.go
@@ -1,0 +1,6 @@
+package entities
+
+type CatFact struct {
+	Id string `json:"id"`
+	Fact string `json:"fact"`
+}
