+++ entities/Error.go
@@ -1,0 +1,6 @@
+package entities
+
+type HTTPError struct {
+	Code int `json:"code"`
+	Message string `json:"message"`
+}
